Rename visual actions from 'peek' to just 'open' screen in modal
